Move calculateFollowupDecisionHand into Rules
